Add middleware for errors and auth, create sign up and sign in, update api endpoints

+++ services/movies.js
@@ -1,4 +1,5 @@
 const MongoLib = require('../lib/mongo');
+const boom = require('@hapi/boom');
 
 class MoviesService {
   constructor() {
@@ -6,8 +7,19 @@ class MoviesService {
     this.mongoDB = new MongoLib();
   }
 
-  async getMovies({ tags }) {
-    const query = tags && { tags: { $in: tags } };
+  async createMovie({ movie, user }) {
+    const createMovieId = await this.mongoDB.create(this.collection, {...movie, user_id_create: user._id, user_email_create: user.email });
+    return createMovieId;
+  }
+
+  async getAllPublicMovies() {
+    let query = { private: false }
+    const movies = await this.mongoDB.getAll(this.collection, query);
+    return movies || [];
+  }
+
+  async getMyMovies({ user }) {
+    const query = { user_id_create: user._id }
     const movies = await this.mongoDB.getAll(this.collection, query);
     return movies || [];
   }
@@ -17,21 +29,32 @@ class MoviesService {
     return movie || {};
   }
 
-  async createMovie({ movie }) {
-    const createMovieId = await this.mongoDB.create(this.collection, movie);
-    return createMovieId;
-  }
+  async updateMovie({ movieId, movie, user } = {}) {
+    let validateMovie = await this.getMovie({ movieId })
 
-  async updateMovie({ movieId, movie } = {}) {
-    const updatedMovieId = await this.mongoDB.update(
-      this.collection,
-      movieId,
-      movie
-    );
+    if (!validateMovie._id) {
+      throw (boom.badRequest(`That movie does not exists "movie_id" ${movieId} `))
+    }
+
+    if (user.email !== validateMovie.user_email_create) {
+      throw (boom.forbidden(`You do not have permission to edit this movie "movie_id" ${movieId} `))
+    }
+
+    const updatedMovieId = await this.mongoDB.update(this.collection, movieId, movie);
     return updatedMovieId;
   }
 
-  async deleteMovie({ movieId }) {
+  async deleteMovie({ movieId, user }) {
+    let validateMovie = await this.getMovie({ movieId });
+
+    if (!validateMovie._id) {
+      throw (boom.badRequest(`That movie does not exists "movie_id" ${movieId} `))
+    }
+
+    if (user.email !== validateMovie.user_email_create) {
+      throw (boom.forbidden(`You do not have permission to delete this movie "movie_id" ${movieId} `))
+    }
+
     const deletedMovieId = await this.mongoDB.delete(this.collection, movieId);
     return deletedMovieId;
   }
